# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fleet Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips (filtered by driver if not admin)
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    put:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: Trip updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /fuelings:
    get:
      operationId: listFuelings
      tags: [fuelings]
      summary: List all fuelings
      responses:
        "200":
          description: List of fuelings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fueling"
    post:
      operationId: createFueling
      tags: [fuelings]
      summary: Create fueling record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuelingRequest"
      responses:
        "201":
          description: Fueling created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fueling"
...
  /maintenance:
    get:
      operationId: listMaintenance
      tags: [maintenance]
      summary: List all maintenance records
      responses:
        "200":
          description: List of maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Maintenance"
    post:
      operationId: createMaintenance
      tags: [maintenance]
      summary: Create maintenance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceRequest"
      responses:
        "201":
          description: Maintenance created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Maintenance"
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierRequest"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierRequest"
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports/fuelings:
    get:
      operationId: getFuelingsReport
      tags: [reports]
      summary: Get fueling report data
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Fueling report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuelingsReport"
  /reports/maintenance:
    get:
      operationId: getMaintenanceReport
      tags: [reports]
      summary: Get maintenance report data
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceReport"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Trip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        origin:
          type: string
        destination:
          type: string
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
          nullable: true
        actualStart:
          type: string
          nullable: true
        actualEnd:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        vehicleId:
          type: integer
          nullable: true
        vehiclePlate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        distance:
          type: number
          nullable: true
        createdAt:
          type: string
      required: [id, title, origin, destination, scheduledStart, status, createdAt]
    CreateTripRequest:
      type: object
      properties:
        title:
          type: string
        origin:
          type: string
        destination:
          type: string
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
          nullable: true
        driverId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        distance:
          type: number
          nullable: true
      required: [title, origin, destination, scheduledStart]
    UpdateTripRequest:
      type: object
      properties:
        title:
          type: string
        origin:
          type: string
        destination:
          type: string
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
          nullable: true
        actualStart:
          type: string
          nullable: true
        actualEnd:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        driverId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        distance:
          type: number
          nullable: true
    Fueling:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
          nullable: true
        driverId:
          type: integer
          nullable: true
        driverName:
          type: string
          nullable: true
        date:
          type: string
        liters:
          type: number
        pricePerLiter:
          type: number
        totalCost:
          type: number
        mileage:
          type: number
        station:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, vehicleId, date, liters, pricePerLiter, totalCost, mileage, createdAt]
    CreateFuelingRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        driverId:
          type: integer
          nullable: true
        date:
          type: string
        liters:
          type: number
        pricePerLiter:
          type: number
        totalCost:
          type: number
        mileage:
          type: number
        station:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [vehicleId, date, liters, pricePerLiter, totalCost, mileage]
...
    Maintenance:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehiclePlate:
          type: string
          nullable: true
        type:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        cost:
          type: number
          nullable: true
        mileage:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        partsUsed:
          type: array
          items:
            $ref: "#/components/schemas/MaintenancePart"
        createdAt:
          type: string
      required: [id, vehicleId, type, description, date, status, createdAt]
    MaintenancePart:
      type: object
      properties:
        inventoryItemId:
          type: integer
        itemName:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
      required: [inventoryItemId, itemName, quantity, unitCost]
    CreateMaintenanceRequest:
      type: object
      properties:
        vehicleId:
          type: integer
        type:
          type: string
        description:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed]
        cost:
          type: number
          nullable: true
        mileage:
          type: number
          nullable: true
        supplierId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        partsUsed:
          type: array
          items:
            $ref: "#/components/schemas/MaintenancePart"
      required: [vehicleId, type, description, date, status]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name]
    UpdateSupplierRequest:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    FuelingsReport:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Fueling"
        totalLiters:
          type: number
        totalCost:
          type: number
        averagePricePerLiter:
          type: number
        byVehicle:
          type: array
          items:
            type: object
            properties:
              vehiclePlate:
                type: string
              totalLiters:
                type: number
              totalCost:
                type: number
            required: [vehiclePlate, totalLiters, totalCost]
      required: [records, totalLiters, totalCost, averagePricePerLiter, byVehicle]
    MaintenanceReport:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/Maintenance"
        totalCost:
          type: number
        byVehicle:
          type: array
          items:
            type: object
            properties:
              vehiclePlate:
                type: string
              totalCost:
                type: number
              count:
                type: integer
            required: [vehiclePlate, totalCost, count]
        byType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
              totalCost:
                type: number
            required: [type, count, totalCost]
      required: [records, totalCost, byVehicle, byType]
...