# lib/api-zod/src/generated/types/updateTripRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fleet Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTripRequestStatus } from "./updateTripRequestStatus";

export interface UpdateTripRequest {
  title?: string;
  origin?: string;
  destination?: string;
  scheduledStart?: string;
  scheduledEnd?: string | null;
  actualStart?: string | null;
  actualEnd?: string | null;
  status?: UpdateTripRequestStatus;
  driverId?: number | null;
  vehicleId?: number | null;
  notes?: string | null;
  distance?: number | null;
}